+++ qa-test/src/bin/embassy_executor_benchmark.rs
@@ -91,9 +91,9 @@ async fn main(spawner: Spawner) {
     esp_rtos::start(systimer.alarm0, sw_int.software_interrupt0);
     println!("Embassy initialized!");
 
-    spawner.spawn(TASK1.spawn(|| Task1 {})).unwrap();
-    spawner.spawn(task2()).unwrap();
-    spawner.spawn(task3()).unwrap();
+    spawner.spawn(TASK1.spawn(|| Task1 {}).unwrap());
+    spawner.spawn(task2().unwrap());
+    spawner.spawn(task3().unwrap());
 
     println!("Starting test");